Move hierarchy traversal to base Elaborator. More Serdes. Remove NoParams from parameter-less generated Module naming
