+++ hdl21/elab/elaborators/generators.py
@@ -2,43 +2,42 @@
 # Generator Elaborator 
 """
 
-# Std-Lib Imports
-import copy
-from typing import Union, Any, Dict, List, Optional, Tuple
-
-# PyPi
-from pydantic.dataclasses import dataclass
-
 # Local imports
-from ...connect import connectable
-from ...module import Module, ExternalModuleCall
-from ...instance import InstArray, Instance, PortRef
-from ...primitives import PrimitiveCall
-from ...bundle import AnonymousBundle, Bundle, BundleInstance, _check_compatible
-from ...signal import PortDir, Signal, Visibility, Slice, Concat, Sliceable, NoConn
-from ...generator import Generator, GeneratorCall, Default as GenDefault
-from ...params import _unique_name
+from ...module import Module
+from ...instance import InstArray, Instance
+from ...generator import GeneratorCall, Default as GenDefault
+from ...params import HasNoParams, _unique_name
 from ...instantiable import Instantiable
 
 # Import the base class
-from .base import _Elaborator
+from .base import Elaborator
 
 
-class GeneratorElaborator(_Elaborator):
-    """ Hierarchical Generator Elaborator
-    Walks a hierarchy from `top` calling Generators. """
+class GeneratorElaborator(Elaborator):
+    """ 
+    # Generator Elaborator
+    
+    Walks a hierarchy from `top` calling Generators. 
+    This pass generally needs to run first among elaboration passes, 
+    or else often times, if `top` is a GeneratorCall, 
+    all other passes really don't have anything to do. 
+    
+    This fact is embedded in a few places throughout `GeneratorElaborator` 
+    and how it inherits behavior from the base `Elaborator`. 
+    Most (probably all) other passes require there be no `GeneratorCall`s 
+    remaining as they operate, and instead act on those Call's `result` properties. 
+    For reuse, this fact is embedded into the base-class in a few places, 
+    and `GeneratorElaborator`'s special-ish case is left to over-ride it. 
+    """
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.generator_calls = dict()  # GeneratorCalls to their (Module) results
-        self.modules = dict()  # Module ids to references
-        self.primitive_calls = dict()  # PrimitiveCall ids to references
-        self.ext_module_calls = dict()  # PrimitiveCall ids to references
 
     def elaborate_top(self):
         """ Elaborate our top node """
         if isinstance(self.top, Module):
-            return self.elaborate_module(self.top)
+            return self.elaborate_module_base(self.top)
         if isinstance(self.top, GeneratorCall):
             return self.elaborate_generator_call(self.top)
         msg = f"Invalid Elaboration top-level {self.top}, must be a Module or Generator"
@@ -89,40 +88,21 @@ class GeneratorElaborator(_Elaborator):
         # Give the GeneratorCall a reference to its result, and store it in our local dict
         call.result = m
         self.generator_calls[call] = m
+
         # Create a unique name
         # If the Module that comes back is anonymous, start by giving it a name equal to the Generator's
         if m.name is None:
             m.name = call.gen.func.__name__
+
         # Then add a unique suffix per its parameter-values
-        m.name += "(" + _unique_name(call.arg) + ")"
+        if not isinstance(call.arg, HasNoParams):
+            m.name += "(" + _unique_name(call.arg) + ")"
+
         # Update the Module's `pymodule`, which generally at this point is `hdl21.generator`
         m._pymodule = call.gen.pymodule
+
         # And elaborate the module
-        return self.elaborate_module(m)
-
-    def elaborate_module(self, module: Module) -> Module:
-        """ Elaborate Module `module`. 
-        Primarily performs flattening of Instance Arrays, 
-        and re-connecting to the resultant flattened instances  """
-        if id(module) in self.modules:  # Already done!
-            return module
-
-        if not module.name:
-            msg = f"Anonymous Module {module} cannot be elaborated (did you forget to name it?)"
-            raise RuntimeError(msg)
-
-        # Depth-first traverse instances, ensuring their targets are defined
-        for inst in module.instances.values():
-            self.elaborate_instance(inst)
-        for arr in module.instarrays.values():
-            self.elaborate_instance_array(arr)
-        # Also visit bundle instances, turning off their pre-elab magic
-        for bundle in module.bundles.values():
-            self.elaborate_bundle_instance(bundle)
-
-        # Store a reference to the now-expanded Module in our cache, and return it
-        self.modules[id(module)] = module
-        return module
+        return self.elaborate_module_base(m)  # Note the `_base` here!
 
     def elaborate_instance(self, inst: Instance) -> Instantiable:
         """ Elaborate a Module Instance. """
